move skeleton code to skeleton folder
