+++ cpp/include/reneu/skeleton/skeleton.hpp
